backend/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.3.0

# Vector database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# Text processing
mecab-python3>=1.0.0
konlpy>=0.6.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web interface
fastapi>=0.104.1,<1.0.0
uvicorn>=0.24.0,<1.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# Development and testing
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<1.0.0

# === 딥러닝 프레임워크 ===
torch>=2.2.0,<3.0.0

# === 한국어 처리 ===
jpype1>=1.4.0,<2.0.0
nltk>=3.8.1,<4.0.0

# === 유틸리티 ===
tqdm>=4.66.1,<5.0.0
colorama>=0.4.6,<1.0.0
psutil>=5.9.0,<6.0.0
loguru>=0.7.2,<1.0.0

# === 데이터 시각화 ===
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0

# === 추가된 패키지 ===
jq>=1.9.1,<2.0.0
transformers>=4.36.0,<5.0.0
huggingface-hub>=0.20.3,<1.0.0 